Refactored uagent, interpreter, and pvt task code
