+++ run.sh
@@ -6,5 +6,5 @@ pkill -f "java"
 echo "Starting ontology server..."
 java -jar lib/fuseki/fuseki-server.jar --update &
 
-echo "Starting agent..."
-python3 uagent.py
+echo "Starting simulation..."
+python3 run.py
