+++ tasks/pvt/__init__.py
@@ -1,2 +1,2 @@
-from .pvt_agent import PVTAgent
-from .pvt_task import PVTTask
+from .agent import PVTAgent
+from .task import PVTTask
