--- ua/__init__.py
@@ -1,1 +1,0 @@
-from .ua import UndifferentiatedAgent
